Refactor Model class into separate module
